+++ apps/client/src/app/pages/public/public-page.component.ts
@@ -1,4 +1,3 @@
-import { DataService } from '@ghostfolio/client/services/data.service';
 import { UNKNOWN_KEY } from '@ghostfolio/common/config';
 import { prettifySymbol } from '@ghostfolio/common/helper';
 import {
@@ -6,8 +5,19 @@ import {
   PublicPortfolioResponse
 } from '@ghostfolio/common/interfaces';
 import { Market } from '@ghostfolio/common/types';
+import { GfHoldingsTableComponent } from '@ghostfolio/ui/holdings-table/holdings-table.component';
+import { GfPortfolioProportionChartComponent } from '@ghostfolio/ui/portfolio-proportion-chart/portfolio-proportion-chart.component';
+import { GfValueComponent } from '@ghostfolio/ui/value';
 
-import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import {
+  ChangeDetectorRef,
+  Component,
+  CUSTOM_ELEMENTS_SCHEMA,
+  OnInit
+} from '@angular/core';
+import { MatButtonModule } from '@angular/material/button';
+import { MatCardModule } from '@angular/material/card';
 import { ActivatedRoute, Router } from '@angular/router';
 import { AssetClass } from '@prisma/client';
 import { StatusCodes } from 'http-status-codes';
@@ -16,14 +26,26 @@ import { DeviceDetectorService } from 'ngx-device-detector';
 import { EMPTY, Subject } from 'rxjs';
 import { catchError, takeUntil } from 'rxjs/operators';
 
+import { GfWorldMapChartComponent } from '../../components/world-map-chart/world-map-chart.component';
+import { DataService } from '../../services/data.service';
+
 @Component({
   host: { class: 'page' },
   selector: 'gf-public-page',
   styleUrls: ['./public-page.scss'],
   templateUrl: './public-page.html',
-  standalone: false
+  imports: [
+    CommonModule,
+    GfHoldingsTableComponent,
+    GfPortfolioProportionChartComponent,
+    GfValueComponent,
+    GfWorldMapChartComponent,
+    MatButtonModule,
+    MatCardModule
+  ],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
-export class PublicPageComponent implements OnInit {
+export class GfPublicPageComponent implements OnInit {
   public continents: {
     [code: string]: { name: string; value: number };
   };