+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactored the benchmark comparator component to standalone
 - Refactored the portfolio summary component to standalone
 - Refactored the world map chart component to standalone
+- Refactored the public page component to standalone
 - Enbabled the trim option in the `extract-i18n` configuration
 - Improved the language localization for German (`de`)
 - Upgraded the _Stripe_ dependencies